inference/requirements.txt
fastapi
uvicorn
mlflow
numpy
python-dotenv